pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pytest_notebook"
dynamic = ["version", "description"]
authors = [{name = "Chris Sewell", email = "chrisj_sewell@hotmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
  "Development Status :: 4 - Beta",
  "Framework :: Pytest",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Testing",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: Implementation :: CPython",
  "Operating System :: OS Independent",
  "License :: OSI Approved :: BSD License",
]
requires-python = "~=3.7"
dependencies = [
  "pytest>=3.5.0",
  "attrs<23,>=19",
  "nbclient~=0.5.10",
  "nbdime",
  "nbformat",
  "jsonschema",
]

[project.urls]
Source = "https://github.com/chrisjsewell/pytest-notebook"
Documentation = "https://pytest-notebook.readthedocs.io"

[project.optional-dependencies]
testing = [
  "pytest-cov",
  "pytest-regressions",
  "ipykernel",
  "coverage~=5.0",
  "black==19.3b0",
  "beautifulsoup4==4.8.0",
]
code_style = [
  "pre-commit~=2.7.0",
  "doc8<0.9.0,>=0.8.0",
]
docs = [
  "sphinx>=3",
  "pyyaml",
  "sphinx-book-theme~=0.1.0",
  "myst-nb~=0.13.1",
  "coverage~=5.0",
]

[project.entry-points."pytest11"]
nb_regression = "pytest_notebook.plugin"

[project.entry-points."nbreg.post_proc"]
coalesce_streams = "pytest_notebook.post_processors:coalesce_streams"
blacken_code = "pytest_notebook.post_processors:blacken_code"
beautifulsoup = "pytest_notebook.post_processors:beautifulsoup"

[tool.isort]
profile = "black"
force_sort_within_sections = true
